todo: Implement ticket assign completion
Here a query is used, that will get the name of the hut user, a list of
already assigned users and a list of assignees of other tickets from the
same tracker. The own username is added to the pool of other assignees
and then every user that is already assigned will be removed from the
pool. The resulting list is the completion.