Fix BSD install error when mixing `-d` and `-C` (via `-p`)
>```
>install: the -d and -C options may not be specified together
>```

This is because of the `-p` flag on the first `install` command line;

>```
>     -p      Preserve the modification time.  Copy the file, as if the -C
>             (compare and copy) option is specified, except if the target file
>             doesn't already exist or is different, then preserve the
>             modification time of the file.
>```

This reads to me like the target file will have the same modification
time as the source file, so at first, I thought we're copying
directories across and we want to preserve the modification times of the
directories, but even after a `make`, the directories being installed
don't exist in the repo directory, they're just being created directly
on the target path. Using the `-p` flag is just pointless in this case.