Fix BSD install error when using `-t` option
>```
>install: -t: No such file or directory
>```

BSD install doesn't have the `-t` option, but according to the GNU info
page for `install`:

>```
>   • If the ‘--target-directory’ (‘-t’) option is given, or failing that
>     if the last file is a directory and the ‘--no-target-directory’
>     (‘-T’) option is not given, ‘install’ copies each SOURCE file to
>     the specified directory, using the SOURCEs’ names.
>```

The directories are strictly specified with a trailing slash, and I
suspect if the directory didn't exist, GNU install would still treat
it as a directory and complain about it being missing. That said, we
don't need to rely on that because the previous line which gives us the
error about combining `-C` and `-d` is exactly the line that creates the
required directories.