git: Change webhook command handling
These changes are mostly based on the todo schema.
The payload URL was switched from the first argument to a flag just like
in the todo ticket webhook creation so that all webhook commands will
have the same syntax/behaviour.
The command itself was changed from webhook to user-webhook because we
can expect repository webhooks in the future. For reference we have
three kinds of webhooks on the todo side right now: User, tracker and
ticket webhooks.