export: add lists.sr.ht support
I filed two follow-up tickets for issues found in lists.sr.ht that
affect this exporter:

https://todo.sr.ht/~sircmpwn/lists.sr.ht/197
https://todo.sr.ht/~sircmpwn/lists.sr.ht/198

I also went with a slightly different approach here which, upon
encountering an error exporting a single list, will skip that list and
continue with the rest, returning the error at the end. This allows it
to keep trying for lists that work and then retry the lists that did not
work on the next attempt, since the export is not registered as
successful and therefore does not write the stamp that would cause
lists.sr.ht to be skipped on the next run.

I used this approach here specifically because of these upstream
lists.sr.ht bugs I found exacerbating the issue, but it may be wise to
apply a similar approach to the other services.