webhooks: Read query after events are parsed
Someone posted a command like this in IRC:
`hut git webhook create $URL -e repo:post-update`
We should validate every flag first and then ask the user to input the
query.
This way is much more convenient because there will be feedback in case
of errors before the effort to write the query is made.