builds secrets: Display type
And if the secret type is a normal file, display the path and mode as
well.
The information is now the same as in the webinterface.

Furthermore the tabwriter was removed, since UUIDs have the same, fixed
length and are automatically aligned. Besides it could cause problems
with multi-line strings.