pager: Fix shutdown
Right now the pager does not get properly closed if there appears an
error inside the pagerify function. The cause is log.Fatal which
will immediately exit the program so that deferred functions (that
would close the pager) will be skipped. This in turn could cause the
terminal to break for some shells.

As a fix the pagerify function is refactored to return an error so that
all calls to log.Fatal can be removed from inside.

The newly introduced pagerDone error is simply used to indicate that the
pager can be closed.