meta update: Improve bio clearing
Instead of introducing a second flag to clear the biography, be a bit
smarter and clear it if the user sets the new bio to an empty string.
Besides the prefill explanation was dropped, since it is only really
useful if we parse two different things (e.g. name and description) from
one file.