pager: Support requesting a number of objects
The GraphQL API does not yet support some kind of "count" filter/parameter.
So for now we let the pager count the objects and abort if either enough
are collected or there are none left.

The actual count implementation for each command is still missing and
will be implemented in follow-up commits.